sweagent-v3/molecule_visualizer/requirements.txt
streamlit==1.31.1
rdkit==2023.9.1
Pillow==10.0.0